+++ client/src/components/Schedule/CalendarView/Calendar.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { ConfigProvider } from 'antd';
-import MobileCalendar from './MobileCalendar';
-import DesktopCalendar from './DesktopCalendar';
 import { isMobile } from 'mobile-device-detect';
 import en_GB from 'antd/lib/locale-provider/en_GB';
-import { ScheduleViewProps } from 'components/Schedule/ScheduleView';
+import { ScheduleViewProps } from '../ScheduleView';
+import MobileCalendar from './MobileCalendar';
+import DesktopCalendar from './DesktopCalendar';
 
 export const CalendarView: React.FC<ScheduleViewProps> = ({ data, courseAlias, settings }) => {
   return (
