--- client/src/components/Schedule/EventModalForm/index.ts
@@ -1,1 +1,0 @@
-export { default as EventModalForm } from './EventModalForm';
