+++ client/src/components/Schedule/ManageEventModalForm/ManageEventModalForm.tsx
@@ -1,24 +1,25 @@
 import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import { Modal, Tabs } from 'antd';
-import FormEntity from './EventForm';
-import useWindowDimensions from '../../../utils/useWindowDimensions';
+import useWindowDimensions from 'utils/useWindowDimensions';
 import { CourseEvent, CourseTaskDetails } from 'services/course';
 import { formatTimezoneToUTC } from 'services/formatter';
-import TaskDetails from '../EventDetails/TaskDetails';
-import EventDetails from '../EventDetails/EventDetails';
+import { ScheduleSettings } from '../useScheduleSettings';
+import { TaskDetails, EventDetails } from '../EventDetails';
+import ManageEventForm from './ManageEventForm';
 
 const { TabPane } = Tabs;
 
-type Props = {
+interface ManageEventModalFormProps {
   visible: boolean;
   handleCancel: () => void;
   courseId: number;
   editableRecord?: CourseEvent | null;
   refreshData: Function;
-};
+  settings: ScheduleSettings;
+}
 
-const EventModalForm: React.FC<Props> = ({ visible, handleCancel, courseId, editableRecord, refreshData }) => {
+const ManageEventModalForm: React.FC<ManageEventModalFormProps> = ({ visible, handleCancel, courseId, editableRecord, refreshData, settings }) => {
   const router = useRouter();
   const { course } = router.query;
   const alias = Array.isArray(course) ? course[0] : course;
@@ -47,7 +48,7 @@ const EventModalForm: React.FC<Props> = ({ visible, handleCancel, courseId, edit
       title={
         <Tabs defaultActiveKey="1">
           <TabPane tab="New entity" key="1">
-            <FormEntity
+            <ManageEventForm
               handleCancel={handleCancel}
               onFieldsChange={onFieldsChange}
               courseId={courseId}
@@ -61,6 +62,7 @@ const EventModalForm: React.FC<Props> = ({ visible, handleCancel, courseId, edit
             {entityType === 'task' && entityData && (
               <TaskDetails
                 taskData={getEntityDataForPreview(entityType, entityData) as CourseTaskDetails}
+                settings={settings}
                 alias={alias}
                 isPreview
                 isAdmin
@@ -69,6 +71,7 @@ const EventModalForm: React.FC<Props> = ({ visible, handleCancel, courseId, edit
             {entityType === 'event' && entityData && (
               <EventDetails
                 eventData={getEntityDataForPreview(entityType, entityData) as CourseEvent}
+                settings={settings}
                 alias={alias}
                 isPreview
                 isAdmin
@@ -106,7 +109,6 @@ const getEntityDataForPreview = (entityType: string, entityData: any) => {
       githubPrRequired: entityData.githubPrRequired,
       sourceGithubRepoUrl: entityData.sourceGithubRepoUrl,
       githubRepoName: entityData.githubRepoName,
-      // attributes: JSON.parse(entityData.attributes ?? '{}'),
       checker: entityData.checker,
       pairsCount: entityData.pairsCount,
     };
@@ -127,4 +129,4 @@ const getEntityDataForPreview = (entityType: string, entityData: any) => {
   };
 };
 
-export default EventModalForm;
+export default ManageEventModalForm;
