+++ client/src/components/Schedule/ManageEventModalForm/index.ts
@@ -1,0 +1,1 @@
+export { default as ManageEventModalForm } from './ManageEventModalForm';
